[Feat] #21 - 뷰와 뷰컨 분리
